dailybuild: ソリューション依存ファイルの追加。 warning の抑止。
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@7524 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Physics/PHContactDetector.cpp
@@ -605,7 +605,7 @@ bool PHContactDetector::Detect(unsigned ct, double dt, int mode, bool continuous
 				}
 				cur.insert(it->index);
 #ifdef _DEBUG
-				if (nMaxOverlapObject < (int)cur.size()) nMaxOverlapObject = cur.size();
+				if (nMaxOverlapObject < (int)cur.size()) nMaxOverlapObject = (int)cur.size();
 #endif
 			}else{
 				cur.erase(it->index);			//	終端なので削除．
